fix dtype

+++ diffuserslib/inference/arch/StableDiffusionPipelines.py
@@ -240,7 +240,7 @@ class StableDiffusionAnimateDiffPipelineWrapper(StableDiffusionPipelineWrapper):
     def __init__(self, params:GenerationParameters, device):
         self.features = self.getPipelineFeatures(params)
         cls = self.getPipelineClass(params)
-        self.adapter = MotionAdapter.from_pretrained("guoyww/animatediff-motion-adapter-v1-5-2", torch_dtype=self.dtype)
+        self.adapter = MotionAdapter.from_pretrained("guoyww/animatediff-motion-adapter-v1-5-2", torch_dtype=torch.float16)
         super().__init__(cls, params, device, dtype = torch.float16)
